# Cache CloudStorage credentials