The credentials token source has built in caching, only generating a token when needed.

Caching the credentials object allows to reuse the same source for the same scopes, leveraging that caching.

Otherwise, CloudStorageFS was generating new tokens every time a method was being called, wasting a significant amount of CPU.