@lavoiesl 2020-09-17
It's doable, but it would make the api a bit harder to use the constructor of the object would need to know ahead of time which scope it would require. It could be a valid, I’m not sure. That being said, it'd be a significantly larger change since it would break the constructor, so perhaps not do it in the scope of this PR.