@lavoiesl 2020-09-18
Yesterday, we ran a load test; I took a second look at our profiler.

For the top 50% weight (to reduce noise):

| Operation                | CPU       | Allocated Heap |
|--------------------------|-----------|----------------|
| Save                     | 335.57 ms | 39.2 MiB       |
| JSON encode              | 97.83 ms  | 17.87 MiB      |
| Create (without writing) | 175.22 ms | 17.91 MiB      |
| bucketHandle             | 150.26 ms | 14.6 MiB       |
| client                   | 137.04 ms | 12.38 MiB      |
| findCredentials          | 135.65 ms | 12.28 MiB      |
| ObjectsInsertCall.doRequest          | 3.83 s    | 126.5 MiB      |
| jwtSource.Token          | 3.59 s    | 126.5 MiB      |

Finding credentials takes about 77% of the "Create" CPU time:
![image](https://user-images.githubusercontent.com/1216046/93598558-d73b1d00-f98a-11ea-81a7-d29df6600ef7.png)

Token generation takes up most of the upload CPU time:
![image](https://user-images.githubusercontent.com/1216046/93598507-c12d5c80-f98a-11ea-81f2-aca401f8d4fe.png)

Similar story for allocated heap:
![image](https://user-images.githubusercontent.com/1216046/93598751-241ef380-f98b-11ea-93f1-a0bbb6f9d1eb.png)
![image](https://user-images.githubusercontent.com/1216046/93598791-36009680-f98b-11ea-8e8d-547f0c4a098a.png)

---

**Conclusion:** Credentials recreation is the culprit and client recreation is negligible