@lavoiesl 2020-09-18
Tested this PR with a benchmark from my computer: https://github.com/Shopify/go-storage/pull/21

```
go test -bench=BenchmarkCloudStorageFS -benchtime=4s -cpuprofile=cpu.out -benchmem -memprofile=mem.out ./...

Before
BenchmarkCloudStorageFS/create-8         	      18	 268202072 ns/op	 8575851 B/op	     857 allocs/op
BenchmarkCloudStorageFS/read-8           	      38	 119780156 ns/op	  149152 B/op	     658 allocs/op

After
BenchmarkCloudStorageFS/create-8         	      24	 192845307 ns/op	 8434111 B/op	     400 allocs/op
BenchmarkCloudStorageFS/read-8           	      70	  67424345 ns/op	   19537 B/op	     211 allocs/op
```

Writes are 28% faster, reads are 44% faster.

Comparing the profiles with `go tool pprof -http 127.0.0.1:1234 -base cpu.before.out cpu.out`, the `Transport.RoundTrip` goes from about 17% of the CPU usage to essentially 0, so I think it demonstrates that while not being ideal, this PR is sufficient to fix the issue.